+++ server/src/main/java/org/allaymc/server/entity/Metadata.java
@@ -5,8 +5,8 @@ import lombok.Getter;
 import org.allaymc.api.block.type.BlockState;
 import org.allaymc.api.entity.data.EntityData;
 import org.allaymc.api.entity.data.EntityFlag;
-import org.allaymc.api.math.MathUtils;
 import org.allaymc.api.registry.Registries;
+import org.allaymc.server.utils.NetworkHelper;
 import org.allaymc.server.utils.ReflectionUtils;
 import org.cloudburstmc.protocol.bedrock.data.definitions.BlockDefinition;
 import org.cloudburstmc.protocol.bedrock.data.entity.EntityDataMap;
@@ -61,8 +61,8 @@ public final class Metadata {
         return switch (data) {
             // Null value is possible here
             case null -> null;
-            case org.cloudburstmc.math.vector.Vector3f v -> MathUtils.toJOMLVec(v);
-            case org.cloudburstmc.math.vector.Vector3i v -> MathUtils.toJOMLVec(v);
+            case org.cloudburstmc.math.vector.Vector3f v -> NetworkHelper.fromNetwork(v);
+            case org.cloudburstmc.math.vector.Vector3i v -> NetworkHelper.fromNetwork(v);
             case BlockDefinition b -> Registries.BLOCK_STATE_PALETTE.get(b.getRuntimeId());
             default -> data;
         };
@@ -70,8 +70,8 @@ public final class Metadata {
 
     private Object toNetwork(Object data) {
         return switch (data) {
-            case Vector3fc v -> MathUtils.toCBVec(v);
-            case Vector3ic v -> MathUtils.toCBVec(v);
+            case Vector3fc v -> NetworkHelper.toNetwork(v);
+            case Vector3ic v -> NetworkHelper.toNetwork(v);
             case BlockState b -> (BlockDefinition) b::blockStateHash;
             case Color c -> c.getRGB();
             default -> data;