+++ CHANGELOG.md
@@ -147,6 +147,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Removed network related methods in classes `ScoreboardLine` and `Scorer`.
 - (API) Removed method `BlockState.toNetworkDefinition()`.
 - (API) Removed method `Container.toNetworkItemData()`.
+- (API) Removed methods `toNetwork()`, `getChunkDataPacket()` and `toEntryInfo()` in class `Pack`.
 
 ## 0.7.1 (API 0.11.0) - 2025/8/20